modularize step 2 - download logic, plus fix throttling
